backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
cloudinary==1.36.0
django-cors-headers==4.3.0
geopy==2.4.1
requests==2.31.0
# twilio==8.10.0  # Optional: Only needed if using Twilio for SMS (pip install twilio)
